messy commit, brute force, late night auth work... but it works! todo: refactor.

+++ app/src/main/java/com/eveningoutpost/dexdrip/G5Model/BondRequestTxMessage.java
@@ -1,16 +1,17 @@
 package com.eveningoutpost.dexdrip.G5Model;
 
 import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
 
 /**
  * Created by joeginley on 3/16/16.
  */
 public class BondRequestTxMessage extends TransmitterMessage {
-    int opcode = 0x7;
+    byte opcode = 0x7;
 
     public BondRequestTxMessage() {
         data = ByteBuffer.allocate(1);
-        data.put((byte)opcode);
+        data.put(opcode);
 
         byteSequence = data.array();
     }
